backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
google-auth==2.35.0
requests==2.32.3
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
httpx
